CA-144231: Fix up the top level Makefile to validate whether it has imported key build variables provided by build.hg.
Previously the clean target could execute 'rm -rf /' if the variable '$(MY_OBJ_DIR)' was not defined. That statement has also been fixed, along with providing a validate target that checks for the existence of '$(MY_OBJ_DIR)' as an indicator of whether build.hg is present.

Long term, we would like to remove all internal dependencies the build process has - but that is not yet possible.

Signed-off-by: Rob Dobson <rob.dobson@citrix.com>