CP-13382: updgrade openssl to 1.0.2a and stunnel to 4.56 in order to support TLSv1.2
Due to security concerns over SSLv3 and previous TLS versions, this
patch upgrades the packages of OpenSSL and Stunnel shipped in the
TransferVM so that stunnel can be configured to support TLSv1.2 only.

This change means that iSCSI initiators using SSL for iSCSI targets
provided by the Transfer VM will need to ensure they are TLSv1.2
capable.

Note that with upgrade to stunnel 4.56 a patch was also taken from Ross
Lagerwall <ross.lagerwall@citrix.com> as consumed in Dom0:

        xen-dist-ocaml.hg/rev/9cba4fda4ad6

Signed-off-by: Rob Dobson <rob.dobson@citrix.com>